% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Make a plot of residuals vs fitted values}
\usage{
plot_residuals(model)
}
\arguments{
\item{model}{an lm object. The function will extract
residuals and fitted values from the model object.}
}
\value{
a ggplot object. A plot showing residual on the
  y-axis and fitted value on the x-axis.
}
\description{
Plot residuals against fitted values to check for 
homoscedasticity.
}
\examples{

model <- lm(mpg~wt, data = mtcars)
plot <- plot_residuals(model)
plot
}
